Remove files for project and add image in title.
