+++ gemma-teixido/challenge/cv-fullstack/frontend/src/App.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function App() {
+  return (
+    <div>
+      <h1>Title CV</h1>
+    </div>
+  );
+}
+
+export default App;
